requirements.txt
Flask==2.2.5
keras==2.12.0
mysql_connector_repackaged==0.3.1
nltk==3.8.1
numpy==1.23.5
